backend/requirements.txt
aiobotocore==2.15.2
aiofiles==24.1.0
aiohttp==3.9.5
aioitertools==0.12.0
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.3
boto3==1.35.36
botocore==1.35.36
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorlog==6.8.2
cryptography==42.0.7
dnspython==2.7.0
email_validator==2.1.1
fastapi==0.110.2
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
idna==3.10
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.5
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.3
uvicorn==0.29.0
wordninja==2.0.0
wrapt==1.17.0
yarl==1.17.1
click==8.1.7
ecdsa==0.19.0
envparse==0.2.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
jose==1.0.0
passlib==1.7.4
pyasn1==0.6.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
sniffio==1.3.1